refactor: update import paths for utility functions and add new utility file
