+++ apps/app/components.json
@@ -13,9 +13,8 @@
   "iconLibrary": "lucide",
   "aliases": {
     "components": "~/components",
-    "utils": "~/lib/utils",
+    "utils": "~/utils/shadcn-ui",
     "ui": "~/components/ui",
-    "lib": "~/lib",
     "hooks": "~/hooks"
   },
   "registries": {}
