+++ apps/app/src/components/ai-elements/actions.tsx
@@ -9,7 +9,7 @@ import {
   TooltipProvider,
   TooltipTrigger,
 } from '~/components/ui/tooltip';
-import { cn } from '~/lib/utils';
+import { cn } from '~/utils/shadcn-ui';
 
 export type ActionsProps = ComponentProps<'div'>;
