+++ apps/app/src/components/ai-elements/conversation.tsx
@@ -7,7 +7,7 @@ import { ArrowDownIcon } from 'lucide-react';
 import { StickToBottom, useStickToBottomContext } from 'use-stick-to-bottom';
 
 import { Button } from '~/components/ui/button';
-import { cn } from '~/lib/utils';
+import { cn } from '~/utils/shadcn-ui';
 
 export type ConversationProps = ComponentProps<'div'> & {
   children: React.ReactNode;
