+++ apps/app/src/components/ai-elements/loader.tsx
@@ -1,6 +1,6 @@
 import type { HTMLAttributes } from 'react';
 
-import { cn } from '~/lib/utils';
+import { cn } from '~/utils/shadcn-ui';
 
 type LoaderIconProps = {
   size?: number;
