+++ apps/app/src/components/ai-elements/message.tsx
@@ -3,7 +3,7 @@ import type { UIMessage } from 'ai';
 import { cva, type VariantProps } from 'class-variance-authority';
 
 import { Avatar, AvatarFallback, AvatarImage } from '~/components/ui/avatar';
-import { cn } from '~/lib/utils';
+import { cn } from '~/utils/shadcn-ui';
 
 export type MessageProps = HTMLAttributes<HTMLDivElement> & {
   from: UIMessage['role'];
