+++ apps/app/src/components/ai-elements/prompt-input.tsx
@@ -75,7 +75,7 @@ import {
   TooltipContent,
   TooltipTrigger,
 } from '~/components/ui/tooltip';
-import { cn } from '~/lib/utils';
+import { cn } from '~/utils/shadcn-ui';
 
 // ============================================================================
 // Provider Context & Types
