+++ apps/app/src/components/ai-elements/shimmer.tsx
@@ -9,7 +9,7 @@ import {
 } from 'react';
 import { motion } from 'motion/react';
 
-import { cn } from '~/lib/utils';
+import { cn } from '~/utils/shadcn-ui';
 
 export type TextShimmerProps = {
   children: string;
