+++ apps/app/src/components/ai-elements/sources.tsx
@@ -8,7 +8,7 @@ import {
   CollapsibleContent,
   CollapsibleTrigger,
 } from '~/components/ui/collapsible';
-import { cn } from '~/lib/utils';
+import { cn } from '~/utils/shadcn-ui';
 
 export type SourcesProps = ComponentProps<'div'>;
