+++ apps/app/src/components/ui/button.tsx
@@ -2,7 +2,7 @@ import type * as React from 'react';
 import { Slot } from '@radix-ui/react-slot';
 import { cva, type VariantProps } from 'class-variance-authority';
 
-import { cn } from '~/lib/utils';
+import { cn } from '~/utils/shadcn-ui';
 
 const buttonVariants = cva(
   // eslint-disable-next-line max-len
