+++ apps/app/src/components/ui/dialog.tsx
@@ -2,7 +2,7 @@ import type * as React from 'react';
 import * as DialogPrimitive from '@radix-ui/react-dialog';
 import { XIcon } from 'lucide-react';
 
-import { cn } from '~/lib/utils';
+import { cn } from '~/utils/shadcn-ui';
 
 function Dialog({
   ...props
