+++ apps/app/src/components/ui/hover-card.tsx
@@ -1,7 +1,7 @@
 import type * as React from 'react';
 import * as HoverCardPrimitive from '@radix-ui/react-hover-card';
 
-import { cn } from '~/lib/utils';
+import { cn } from '~/utils/shadcn-ui';
 
 function HoverCard({
   ...props
