+++ apps/app/src/components/ui/input-group.tsx
@@ -6,7 +6,7 @@ import { cva, type VariantProps } from 'class-variance-authority';
 import { Button } from '~/components/ui/button';
 import { Input } from '~/components/ui/input';
 import { Textarea } from '~/components/ui/textarea';
-import { cn } from '~/lib/utils';
+import { cn } from '~/utils/shadcn-ui';
 
 function InputGroup({ className, ...props }: React.ComponentProps<'div'>) {
   return (
