+++ apps/app/src/components/ui/input.tsx
@@ -1,6 +1,6 @@
 import type * as React from 'react';
 
-import { cn } from '~/lib/utils';
+import { cn } from '~/utils/shadcn-ui';
 
 function Input({ className, type, ...props }: React.ComponentProps<'input'>) {
   return (
