+++ apps/app/src/components/ui/tooltip.tsx
@@ -1,7 +1,7 @@
 import type * as React from 'react';
 import * as TooltipPrimitive from '@radix-ui/react-tooltip';
 
-import { cn } from '~/lib/utils';
+import { cn } from '~/utils/shadcn-ui';
 
 function TooltipProvider({
   delayDuration = 0,
